backend/requirements.txt
Flask>=2.3.0,<3.0.0 # Latest stable as of writing, check PyPI
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Migrate>=4.0.0,<5.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
Werkzeug>=2.3.0,<3.0.0 # For password hashing
Flask-JWT-Extended>=4.5.0,<5.0.0
gunicorn>=21.0.0,<22.0.0
google-cloud-bigquery>=3.10.0,<4.0.0
google-auth>=2.20.0,<3.0.0 # Often a dependency of google-cloud libraries
PyJWT>=2.7.0,<3.0.0 # Dependency for Flask-JWT-Extended
# uuid is a built-in module